py/requirements.txt
dbus-python~=1.2.18
PyGObject~=3.42.2
